% Generated by build script
\name{dv_dataset}
\alias{dv_dataset}
\title{Get dataset details by DOI}
\usage{
dv_dataset()
}
\arguments{
\item{doi}{Dataset persistent ID (e.g. "doi:10.7910/DVN/EXAMPLE")}
}
\value{tibble: name, doi, description, publisher, published_at,}
\description{Get dataset details by DOI}
