% Generated by build script
\name{dv_dataverse}
\alias{dv_dataverse}
\title{Get info about a specific dataverse collection}
\usage{
dv_dataverse()
}
\arguments{
\item{alias}{Dataverse alias (e.g. "harvard", "iqss")}
}
\value{tibble: id, alias, name, description, creation_date}
\description{Get info about a specific dataverse collection}
